requirements.txt
langmem>=0.0.7
langgraph>=0.0.24
openai>=1.7.0
elevenlabs>=0.2.26
sounddevice>=0.4.6
scipy>=1.11.3
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.3
langchain-core>=0.1.0
langchain-openai>=0.0.5
uuid>=0.1.0
